requirements.txt
numpy==1.23.5
xgboost
Flask==2.2.2
scikit-learn==1.1.3
pandas==1.4.2
matplotlib==3.5.1
notebook==6.5.2
python-dotenv